refactor(config): split schema monolith
