+++ src/config/schema/mod.rs
@@ -13,6 +13,10 @@ use tokio::fs::File;
 use tokio::fs::{self, OpenOptions};
 use tokio::io::AsyncWriteExt;
 
+mod config_impl;
+mod ops;
+mod root_config;
+
 const SUPPORTED_PROXY_SERVICE_KEYS: &[&str] = &[
     "provider.anthropic",
     "provider.compatible",
@@ -59,530 +63,14 @@ static RUNTIME_PROXY_CLIENT_CACHE: OnceLock<RwLock<HashMap<String, reqwest::Clie
 
 // ── Top-level config ──────────────────────────────────────────────
 
-/// Top-level R.A.I.N. configuration, loaded from `config.toml`.
-///
-/// Resolution order: `rain_WORKSPACE` env → `active_workspace.toml` marker → `~/.R.A.I.N./config.toml`.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct Config {
-    /// Workspace directory - computed from home, not serialized
-    #[serde(skip)]
-    pub workspace_dir: PathBuf,
-    /// Path to config.toml - computed from home, not serialized
-    #[serde(skip)]
-    pub config_path: PathBuf,
-    /// API key for the selected provider. Overridden by `rain_API_KEY` or `API_KEY` env vars.
-    pub api_key: Option<String>,
-    /// Base URL override for provider API (e.g. "http://10.0.0.1:11434" for remote Ollama)
-    pub api_url: Option<String>,
-    /// Custom API path suffix for OpenAI-compatible / custom providers
-    /// (e.g. "/v2/generate" instead of the default "/v1/chat/completions").
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub api_path: Option<String>,
-    /// Default provider ID or alias (e.g. `"openrouter"`, `"ollama"`, `"anthropic"`). Default: `"openrouter"`.
-    #[serde(alias = "model_provider")]
-    pub default_provider: Option<String>,
-    /// Default model routed through the selected provider (e.g. `"anthropic/claude-sonnet-4-6"`).
-    #[serde(alias = "model")]
-    pub default_model: Option<String>,
-    /// Optional named provider profiles keyed by id (Codex app-server compatible layout).
-    #[serde(default)]
-    pub model_providers: HashMap<String, ModelProviderConfig>,
-    /// Default model temperature (0.0–2.0). Default: `0.7`.
-    #[serde(
-        default = "default_temperature",
-        deserialize_with = "deserialize_temperature"
-    )]
-    pub default_temperature: f64,
-
-    /// HTTP request timeout in seconds for LLM provider API calls. Default: `120`.
-    ///
-    /// Increase for slower backends (e.g., llama.cpp on constrained hardware)
-    /// that need more time processing large contexts.
-    #[serde(default = "default_provider_timeout_secs")]
-    pub provider_timeout_secs: u64,
-
-    /// Extra HTTP headers to include in LLM provider API requests.
-    ///
-    /// Some providers require specific headers (e.g., `User-Agent`, `HTTP-Referer`,
-    /// `X-Title`) for request routing or policy enforcement. Headers defined here
-    /// augment (and override) the program's default headers.
-    ///
-    /// Can also be set via `rain_EXTRA_HEADERS` environment variable using
-    /// the format `Key:Value,Key2:Value2`. Env var headers override config file headers.
-    #[serde(default)]
-    pub extra_headers: HashMap<String, String>,
-
-    /// Observability backend configuration (`[observability]`).
-    #[serde(default)]
-    pub observability: ObservabilityConfig,
-
-    /// Autonomy and security policy configuration (`[autonomy]`).
-    #[serde(default)]
-    pub autonomy: AutonomyConfig,
-
-    /// Security subsystem configuration (`[security]`).
-    #[serde(default)]
-    pub security: SecurityConfig,
-
-    /// Backup tool configuration (`[backup]`).
-    #[serde(default)]
-    pub backup: BackupConfig,
-
-    /// Data retention and purge configuration (`[data_retention]`).
-    #[serde(default)]
-    pub data_retention: DataRetentionConfig,
-
-    /// Cloud transformation accelerator configuration (`[cloud_ops]`).
-    #[serde(default)]
-    pub cloud_ops: CloudOpsConfig,
-
-    /// Conversational AI agent builder configuration (`[conversational_ai]`).
-    ///
-    /// Experimental / future feature — not yet wired into the agent runtime.
-    /// Omitted from generated config files when disabled (the default).
-    /// Existing configs that already contain this section will continue to
-    /// deserialize correctly thanks to `#[serde(default)]`.
-    #[serde(default, skip_serializing_if = "ConversationalAiConfig::is_disabled")]
-    pub conversational_ai: ConversationalAiConfig,
-
-    /// Managed cybersecurity service configuration (`[security_ops]`).
-    #[serde(default)]
-    pub security_ops: SecurityOpsConfig,
-
-    /// Runtime adapter configuration (`[runtime]`). Controls native vs Docker execution.
-    #[serde(default)]
-    pub runtime: RuntimeConfig,
-
-    /// Reliability settings: retries, fallback providers, backoff (`[reliability]`).
-    #[serde(default)]
-    pub reliability: ReliabilityConfig,
-
-    /// Scheduler configuration for periodic task execution (`[scheduler]`).
-    #[serde(default)]
-    pub scheduler: SchedulerConfig,
-
-    /// Agent orchestration settings (`[agent]`).
-    #[serde(default)]
-    pub agent: AgentConfig,
-
-    /// Pacing controls for slow/local LLM workloads (`[pacing]`).
-    #[serde(default)]
-    pub pacing: PacingConfig,
-
-    /// Skills loading and community repository behavior (`[skills]`).
-    #[serde(default)]
-    pub skills: SkillsConfig,
-
-    /// Model routing rules — route `hint:<name>` to specific provider+model combos.
-    #[serde(default)]
-    pub model_routes: Vec<ModelRouteConfig>,
-
-    /// Embedding routing rules — route `hint:<name>` to specific provider+model combos.
-    #[serde(default)]
-    pub embedding_routes: Vec<EmbeddingRouteConfig>,
-
-    /// Automatic query classification — maps user messages to model hints.
-    #[serde(default)]
-    pub query_classification: QueryClassificationConfig,
-
-    /// Heartbeat configuration for periodic health pings (`[heartbeat]`).
-    #[serde(default)]
-    pub heartbeat: HeartbeatConfig,
-
-    /// Cron job configuration (`[cron]`).
-    #[serde(default)]
-    pub cron: CronConfig,
-
-    /// Channel configurations: Telegram, Discord, Slack, etc. (`[channels_config]`).
-    #[serde(default)]
-    pub channels_config: ChannelsConfig,
-
-    /// Memory backend configuration: sqlite, markdown, embeddings (`[memory]`).
-    #[serde(default)]
-    pub memory: MemoryConfig,
-
-    /// Persistent storage provider configuration (`[storage]`).
-    #[serde(default)]
-    pub storage: StorageConfig,
-
-    /// Tunnel configuration for exposing the gateway publicly (`[tunnel]`).
-    #[serde(default)]
-    pub tunnel: TunnelConfig,
-
-    /// Gateway server configuration: host, port, pairing, rate limits (`[gateway]`).
-    #[serde(default)]
-    pub gateway: GatewayConfig,
-
-    /// Composio managed OAuth tools integration (`[composio]`).
-    #[serde(default)]
-    pub composio: ComposioConfig,
-
-    /// Microsoft 365 Graph API integration (`[microsoft365]`).
-    #[serde(default)]
-    pub microsoft365: Microsoft365Config,
-
-    /// Secrets encryption configuration (`[secrets]`).
-    #[serde(default)]
-    pub secrets: SecretsConfig,
-
-    /// Browser automation configuration (`[browser]`).
-    #[serde(default)]
-    pub browser: BrowserConfig,
-
-    /// Browser delegation configuration (`[browser_delegate]`).
-    ///
-    /// Delegates browser-based tasks to a browser-capable CLI subprocess (e.g.
-    /// Claude Code with `claude-in-chrome` MCP tools). Useful for interacting
-    /// with corporate web apps (Teams, Outlook, Jira, Confluence) that lack
-    /// direct API access. A persistent Chrome profile can be configured so SSO
-    /// sessions survive across invocations.
-    ///
-    /// Fields:
-    /// - `enabled` (`bool`, default `false`) — enable the browser delegation tool.
-    /// - `cli_binary` (`String`, default `"claude"`) — CLI binary to spawn for browser tasks.
-    /// - `chrome_profile_dir` (`String`, default `""`) — Chrome user-data directory for
-    ///   persistent SSO sessions. When empty, a fresh profile is used each invocation.
-    /// - `allowed_domains` (`Vec<String>`, default `[]`) — allowlist of domains the browser
-    ///   may navigate to. Empty means all non-blocked domains are permitted.
-    /// - `blocked_domains` (`Vec<String>`, default `[]`) — denylist of domains. Blocked
-    ///   domains take precedence over allowed domains.
-    /// - `task_timeout_secs` (`u64`, default `120`) — per-task timeout in seconds.
-    ///
-    /// Compatibility: additive and disabled by default; existing configs remain valid when omitted.
-    /// Rollback/migration: remove `[browser_delegate]` or keep `enabled = false` to disable.
-    #[serde(default)]
-    pub browser_delegate: crate::tools::browser_delegate::BrowserDelegateConfig,
-
-    /// HTTP request tool configuration (`[http_request]`).
-    #[serde(default)]
-    pub http_request: HttpRequestConfig,
-
-    /// Multimodal (image) handling configuration (`[multimodal]`).
-    #[serde(default)]
-    pub multimodal: MultimodalConfig,
-
-    /// Web fetch tool configuration (`[web_fetch]`).
-    #[serde(default)]
-    pub web_fetch: WebFetchConfig,
-
-    /// Text browser tool configuration (`[text_browser]`).
-    #[serde(default)]
-    pub text_browser: TextBrowserConfig,
-
-    /// Web search tool configuration (`[web_search]`).
-    #[serde(default)]
-    pub web_search: WebSearchConfig,
-
-    /// Project delivery intelligence configuration (`[project_intel]`).
-    #[serde(default)]
-    pub project_intel: ProjectIntelConfig,
-
-    /// Google Workspace CLI (`gws`) tool configuration (`[google_workspace]`).
-    #[serde(default)]
-    pub google_workspace: GoogleWorkspaceConfig,
-
-    /// Proxy configuration for outbound HTTP/HTTPS/SOCKS5 traffic (`[proxy]`).
-    #[serde(default)]
-    pub proxy: ProxyConfig,
-
-    /// Identity format configuration: OpenClaw or AIEOS (`[identity]`).
-    #[serde(default)]
-    pub identity: IdentityConfig,
-
-    /// Cost tracking and budget enforcement configuration (`[cost]`).
-    #[serde(default)]
-    pub cost: CostConfig,
-
-    /// Peripheral board configuration for hardware integration (`[peripherals]`).
-    #[serde(default)]
-    pub peripherals: PeripheralsConfig,
-
-    /// Delegate tool global default configuration (`[delegate]`).
-    #[serde(default)]
-    pub delegate: DelegateToolConfig,
-
-    /// Delegate agent configurations for multi-agent workflows.
-    #[serde(default)]
-    pub agents: HashMap<String, DelegateAgentConfig>,
-
-    /// Swarm configurations for multi-agent orchestration.
-    #[serde(default)]
-    pub swarms: HashMap<String, SwarmConfig>,
-
-    /// Hooks configuration (lifecycle hooks and built-in hook toggles).
-    #[serde(default)]
-    pub hooks: HooksConfig,
-
-    /// Hardware configuration (wizard-driven physical world setup).
-    #[serde(default)]
-    pub hardware: HardwareConfig,
-
-    /// Voice transcription configuration (Whisper API via Groq).
-    #[serde(default)]
-    pub transcription: TranscriptionConfig,
-
-    /// Text-to-Speech configuration (`[tts]`).
-    #[serde(default)]
-    pub tts: TtsConfig,
-
-    /// External MCP server connections (`[mcp]`).
-    #[serde(default, alias = "mcpServers")]
-    pub mcp: McpConfig,
-
-    /// Dynamic node discovery configuration (`[nodes]`).
-    #[serde(default)]
-    pub nodes: NodesConfig,
-
-    /// Multi-client workspace isolation configuration (`[workspace]`).
-    #[serde(default)]
-    pub workspace: WorkspaceConfig,
-
-    /// Notion integration configuration (`[notion]`).
-    #[serde(default)]
-    pub notion: NotionConfig,
-
-    /// Jira integration configuration (`[jira]`).
-    #[serde(default)]
-    pub jira: JiraConfig,
-
-    /// Secure inter-node transport configuration (`[node_transport]`).
-    #[serde(default)]
-    pub node_transport: NodeTransportConfig,
-
-    /// Knowledge graph configuration (`[knowledge]`).
-    #[serde(default)]
-    pub knowledge: KnowledgeConfig,
-
-    /// LinkedIn integration configuration (`[linkedin]`).
-    #[serde(default)]
-    pub linkedin: LinkedInConfig,
-
-    /// Plugin system configuration (`[plugins]`).
-    #[serde(default)]
-    pub plugins: PluginsConfig,
-
-    /// Locale for tool descriptions (e.g. `"en"`, `"zh-CN"`).
-    ///
-    /// When set, tool descriptions shown in system prompts are loaded from
-    /// `tool_descriptions/<locale>.toml`. Falls back to English, then to
-    /// hardcoded descriptions.
-    ///
-    /// If omitted or empty, the locale is auto-detected from `rain_LOCALE`,
-    /// `LANG`, or `LC_ALL` environment variables (defaulting to `"en"`).
-    #[serde(default)]
-    pub locale: Option<String>,
-
-    /// Verifiable Intent (VI) credential verification and issuance (`[verifiable_intent]`).
-    #[serde(default)]
-    pub verifiable_intent: VerifiableIntentConfig,
-
-    /// Claude Code tool configuration (`[claude_code]`).
-    #[serde(default)]
-    pub claude_code: ClaudeCodeConfig,
-}
-
-/// Multi-client workspace isolation configuration.
-///
-/// When enabled, each client engagement gets an isolated workspace with
-/// separate memory, audit, secrets, and tool restrictions.
-#[allow(clippy::struct_excessive_bools)]
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct WorkspaceConfig {
-    /// Enable workspace isolation. Default: false.
-    #[serde(default)]
-    pub enabled: bool,
-    /// Currently active workspace name.
-    #[serde(default)]
-    pub active_workspace: Option<String>,
-    /// Base directory for workspace profiles.
-    #[serde(default = "default_workspaces_dir")]
-    pub workspaces_dir: String,
-    /// Isolate memory databases per workspace. Default: true.
-    #[serde(default = "default_true")]
-    pub isolate_memory: bool,
-    /// Isolate secrets namespaces per workspace. Default: true.
-    #[serde(default = "default_true")]
-    pub isolate_secrets: bool,
-    /// Isolate audit logs per workspace. Default: true.
-    #[serde(default = "default_true")]
-    pub isolate_audit: bool,
-    /// Allow searching across workspaces. Default: false (security).
-    #[serde(default)]
-    pub cross_workspace_search: bool,
-}
-
-fn default_workspaces_dir() -> String {
-    "~/.R.A.I.N./workspaces".to_string()
-}
-
-impl Default for WorkspaceConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            active_workspace: None,
-            workspaces_dir: default_workspaces_dir(),
-            isolate_memory: true,
-            isolate_secrets: true,
-            isolate_audit: true,
-            cross_workspace_search: false,
-        }
-    }
-}
-
-/// Named provider profile definition compatible with Codex app-server style config.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, Default)]
-pub struct ModelProviderConfig {
-    /// Optional provider type/name override (e.g. "openai", "openai-codex", or custom profile id).
-    #[serde(default)]
-    pub name: Option<String>,
-    /// Optional base URL for OpenAI-compatible endpoints.
-    #[serde(default)]
-    pub base_url: Option<String>,
-    /// Optional custom API path suffix (e.g. "/v2/generate" instead of the
-    /// default "/v1/chat/completions"). Only used by OpenAI-compatible / custom providers.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub api_path: Option<String>,
-    /// Provider protocol variant ("responses" or "chat_completions").
-    #[serde(default)]
-    pub wire_api: Option<String>,
-    /// If true, load OpenAI auth material (OPENAI_API_KEY or ~/.codex/auth.json).
-    #[serde(default)]
-    pub requires_openai_auth: bool,
-    /// Azure OpenAI resource name (e.g. "my-resource" in https://my-resource.openai.azure.com).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub azure_openai_resource: Option<String>,
-    /// Azure OpenAI deployment name (e.g. "gpt-4o").
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub azure_openai_deployment: Option<String>,
-    /// Azure OpenAI API version (defaults to "2024-08-01-preview").
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub azure_openai_api_version: Option<String>,
-}
-
-// ── Delegate Tool Configuration ─────────────────────────────────
-
-/// Global delegate tool configuration for default timeout values.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct DelegateToolConfig {
-    /// Default timeout in seconds for non-agentic sub-agent provider calls.
-    /// Can be overridden per-agent in `[agents.<name>]` config.
-    /// Default: 120 seconds.
-    #[serde(default = "default_delegate_timeout_secs")]
-    pub timeout_secs: u64,
-    /// Default timeout in seconds for agentic sub-agent runs.
-    /// Can be overridden per-agent in `[agents.<name>]` config.
-    /// Default: 300 seconds.
-    #[serde(default = "default_delegate_agentic_timeout_secs")]
-    pub agentic_timeout_secs: u64,
-}
-
-impl Default for DelegateToolConfig {
-    fn default() -> Self {
-        Self {
-            timeout_secs: DEFAULT_DELEGATE_TIMEOUT_SECS,
-            agentic_timeout_secs: DEFAULT_DELEGATE_AGENTIC_TIMEOUT_SECS,
-        }
-    }
-}
-
-// ── Delegate Agents ──────────────────────────────────────────────
-
-/// Configuration for a delegate sub-agent used by the `delegate` tool.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct DelegateAgentConfig {
-    /// Provider name (e.g. "ollama", "openrouter", "anthropic")
-    pub provider: String,
-    /// Model name
-    pub model: String,
-    /// Optional system prompt for the sub-agent
-    #[serde(default)]
-    pub system_prompt: Option<String>,
-    /// Optional API key override
-    #[serde(default)]
-    pub api_key: Option<String>,
-    /// Temperature override
-    #[serde(default)]
-    pub temperature: Option<f64>,
-    /// Max recursion depth for nested delegation
-    #[serde(default = "default_max_depth")]
-    pub max_depth: u32,
-    /// Enable agentic sub-agent mode (multi-turn tool-call loop).
-    #[serde(default)]
-    pub agentic: bool,
-    /// Allowlist of tool names available to the sub-agent in agentic mode.
-    #[serde(default)]
-    pub allowed_tools: Vec<String>,
-    /// Maximum tool-call iterations in agentic mode.
-    #[serde(default = "default_max_tool_iterations")]
-    pub max_iterations: usize,
-    /// Optional timeout in seconds for non-agentic sub-agent provider calls.
-    /// When `None`, falls back to `[delegate].timeout_secs` (default: 120).
-    #[serde(default)]
-    pub timeout_secs: Option<u64>,
-    /// Optional timeout in seconds for agentic sub-agent runs.
-    /// When `None`, falls back to `[delegate].agentic_timeout_secs` (default: 300).
-    #[serde(default)]
-    pub agentic_timeout_secs: Option<u64>,
-    /// Optional skills directory path (relative to workspace root) for scoped skill loading.
-    /// When unset or empty, the sub-agent falls back to the default workspace `skills/` directory.
-    #[serde(default)]
-    pub skills_directory: Option<String>,
-}
-
-fn default_delegate_timeout_secs() -> u64 {
-    DEFAULT_DELEGATE_TIMEOUT_SECS
-}
-
-fn default_delegate_agentic_timeout_secs() -> u64 {
-    DEFAULT_DELEGATE_AGENTIC_TIMEOUT_SECS
-}
-
-// ── Swarms ──────────────────────────────────────────────────────
-
-/// Orchestration strategy for a swarm of agents.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum SwarmStrategy {
-    /// Run agents sequentially; each agent's output feeds into the next.
-    Sequential,
-    /// Run agents in parallel; collect all outputs.
-    Parallel,
-    /// Use the LLM to pick the best agent for the task.
-    Router,
-}
-
-/// Configuration for a swarm of coordinated agents.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct SwarmConfig {
-    /// Ordered list of agent names (must reference keys in `agents`).
-    pub agents: Vec<String>,
-    /// Orchestration strategy.
-    pub strategy: SwarmStrategy,
-    /// System prompt for router strategy (used to pick the best agent).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub router_prompt: Option<String>,
-    /// Optional description shown to the LLM when choosing swarms.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    /// Maximum total timeout for the swarm execution in seconds.
-    #[serde(default = "default_swarm_timeout_secs")]
-    pub timeout_secs: u64,
-}
-
-const DEFAULT_SWARM_TIMEOUT_SECS: u64 = 300;
-
-fn default_swarm_timeout_secs() -> u64 {
-    DEFAULT_SWARM_TIMEOUT_SECS
-}
-
-/// Valid temperature range for all paths (config, CLI, env override).
-pub const TEMPERATURE_RANGE: std::ops::RangeInclusive<f64> = 0.0..=2.0;
-
-/// Default temperature when the field is absent from config.
-const DEFAULT_TEMPERATURE: f64 = 0.7;
+pub use ops::{CloudOpsConfig, ConversationalAiConfig, SecurityOpsConfig};
+pub use root_config::{
+    Config, DelegateAgentConfig, DelegateToolConfig, ModelProviderConfig, SwarmConfig,
+    SwarmStrategy, WorkspaceConfig, TEMPERATURE_RANGE,
+};
 
 fn default_temperature() -> f64 {
-    DEFAULT_TEMPERATURE
+    root_config::DEFAULT_TEMPERATURE
 }
 
 /// Default provider HTTP request timeout: 120 seconds.
@@ -6499,328 +5987,7 @@ impl Default for JiraConfig {
     }
 }
 
-///
-/// Controls the read-only cloud transformation analysis tools:
-/// IaC review, migration assessment, cost analysis, and architecture review.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct CloudOpsConfig {
-    /// Enable cloud operations tools. Default: false.
-    #[serde(default)]
-    pub enabled: bool,
-    /// Default cloud provider for analysis context. Default: "aws".
-    #[serde(default = "default_cloud_ops_cloud")]
-    pub default_cloud: String,
-    /// Supported cloud providers. Default: [`aws`, `azure`, `gcp`].
-    #[serde(default = "default_cloud_ops_supported_clouds")]
-    pub supported_clouds: Vec<String>,
-    /// Supported IaC tools for review. Default: [`terraform`].
-    #[serde(default = "default_cloud_ops_iac_tools")]
-    pub iac_tools: Vec<String>,
-    /// Monthly USD threshold to flag cost items. Default: 100.0.
-    #[serde(default = "default_cloud_ops_cost_threshold")]
-    pub cost_threshold_monthly_usd: f64,
-    /// Well-Architected Frameworks to check against. Default: [`aws-waf`].
-    #[serde(default = "default_cloud_ops_waf")]
-    pub well_architected_frameworks: Vec<String>,
-}
-
-impl Default for CloudOpsConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            default_cloud: default_cloud_ops_cloud(),
-            supported_clouds: default_cloud_ops_supported_clouds(),
-            iac_tools: default_cloud_ops_iac_tools(),
-            cost_threshold_monthly_usd: default_cloud_ops_cost_threshold(),
-            well_architected_frameworks: default_cloud_ops_waf(),
-        }
-    }
-}
-
-impl CloudOpsConfig {
-    pub fn validate(&self) -> Result<()> {
-        if self.enabled {
-            if self.default_cloud.trim().is_empty() {
-                anyhow::bail!(
-                    "cloud_ops.default_cloud must not be empty when cloud_ops is enabled"
-                );
-            }
-            if self.supported_clouds.is_empty() {
-                anyhow::bail!(
-                    "cloud_ops.supported_clouds must not be empty when cloud_ops is enabled"
-                );
-            }
-            for (i, cloud) in self.supported_clouds.iter().enumerate() {
-                if cloud.trim().is_empty() {
-                    anyhow::bail!("cloud_ops.supported_clouds[{i}] must not be empty");
-                }
-            }
-            if !self.supported_clouds.contains(&self.default_cloud) {
-                anyhow::bail!(
-                    "cloud_ops.default_cloud '{}' is not in cloud_ops.supported_clouds {:?}",
-                    self.default_cloud,
-                    self.supported_clouds
-                );
-            }
-            if self.cost_threshold_monthly_usd < 0.0 {
-                anyhow::bail!(
-                    "cloud_ops.cost_threshold_monthly_usd must be non-negative, got {}",
-                    self.cost_threshold_monthly_usd
-                );
-            }
-            if self.iac_tools.is_empty() {
-                anyhow::bail!("cloud_ops.iac_tools must not be empty when cloud_ops is enabled");
-            }
-        }
-        Ok(())
-    }
-}
-
-fn default_cloud_ops_cloud() -> String {
-    "aws".into()
-}
-
-fn default_cloud_ops_supported_clouds() -> Vec<String> {
-    vec!["aws".into(), "azure".into(), "gcp".into()]
-}
-
-fn default_cloud_ops_iac_tools() -> Vec<String> {
-    vec!["terraform".into()]
-}
-
-fn default_cloud_ops_cost_threshold() -> f64 {
-    100.0
-}
-
-fn default_cloud_ops_waf() -> Vec<String> {
-    vec!["aws-waf".into()]
-}
-
-// ── Conversational AI ──────────────────────────────────────────────
-
-fn default_conversational_ai_language() -> String {
-    "en".into()
-}
-
-fn default_conversational_ai_supported_languages() -> Vec<String> {
-    vec!["en".into(), "de".into(), "fr".into(), "it".into()]
-}
-
-fn default_conversational_ai_escalation_threshold() -> f64 {
-    0.3
-}
-
-fn default_conversational_ai_max_turns() -> usize {
-    50
-}
-
-fn default_conversational_ai_timeout_secs() -> u64 {
-    1800
-}
-
-/// Conversational AI agent builder configuration (`[conversational_ai]` section).
-///
-/// **Status: Reserved for future use.** This configuration is parsed but not yet
-/// consumed by the runtime. Setting `enabled = true` will produce a startup warning.
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct ConversationalAiConfig {
-    /// Enable conversational AI features. Default: false.
-    #[serde(default)]
-    pub enabled: bool,
-    /// Default language for conversations (BCP-47 tag). Default: "en".
-    #[serde(default = "default_conversational_ai_language")]
-    pub default_language: String,
-    /// Supported languages for conversations. Default: [`en`, `de`, `fr`, `it`].
-    #[serde(default = "default_conversational_ai_supported_languages")]
-    pub supported_languages: Vec<String>,
-    /// Automatically detect user language from message content. Default: true.
-    #[serde(default = "default_true")]
-    pub auto_detect_language: bool,
-    /// Intent confidence below this threshold triggers escalation. Default: 0.3.
-    #[serde(default = "default_conversational_ai_escalation_threshold")]
-    pub escalation_confidence_threshold: f64,
-    /// Maximum conversation turns before auto-ending. Default: 50.
-    #[serde(default = "default_conversational_ai_max_turns")]
-    pub max_conversation_turns: usize,
-    /// Conversation timeout in seconds (inactivity). Default: 1800.
-    #[serde(default = "default_conversational_ai_timeout_secs")]
-    pub conversation_timeout_secs: u64,
-    /// Enable conversation analytics tracking. Default: false (privacy-by-default).
-    #[serde(default)]
-    pub analytics_enabled: bool,
-    /// Optional tool name for RAG-based knowledge base lookup during conversations.
-    #[serde(default)]
-    pub knowledge_base_tool: Option<String>,
-}
-
-impl ConversationalAiConfig {
-    /// Returns `true` when the feature is disabled (the default).
-    ///
-    /// Used by `#[serde(skip_serializing_if)]` to omit the entire
-    /// `[conversational_ai]` section from newly-generated config files,
-    /// avoiding user confusion over an undocumented / experimental section.
-    pub fn is_disabled(&self) -> bool {
-        !self.enabled
-    }
-}
-
-impl Default for ConversationalAiConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            default_language: default_conversational_ai_language(),
-            supported_languages: default_conversational_ai_supported_languages(),
-            auto_detect_language: true,
-            escalation_confidence_threshold: default_conversational_ai_escalation_threshold(),
-            max_conversation_turns: default_conversational_ai_max_turns(),
-            conversation_timeout_secs: default_conversational_ai_timeout_secs(),
-            analytics_enabled: false,
-            knowledge_base_tool: None,
-        }
-    }
-}
-
-// ── Security ops config ─────────────────────────────────────────
-
-/// Managed Cybersecurity Service (MCSS) dashboard agent configuration (`[security_ops]`).
-#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
-pub struct SecurityOpsConfig {
-    /// Enable security operations tools.
-    #[serde(default)]
-    pub enabled: bool,
-    /// Directory containing incident response playbook definitions (JSON).
-    #[serde(default = "default_playbooks_dir")]
-    pub playbooks_dir: String,
-    /// Automatically triage incoming alerts without user prompt.
-    #[serde(default)]
-    pub auto_triage: bool,
-    /// Require human approval before executing playbook actions.
-    #[serde(default = "default_require_approval")]
-    pub require_approval_for_actions: bool,
-    /// Maximum severity level that can be auto-remediated without approval.
-    /// One of: "low", "medium", "high", "critical". Default: "low".
-    #[serde(default = "default_max_auto_severity")]
-    pub max_auto_severity: String,
-    /// Directory for generated security reports.
-    #[serde(default = "default_report_output_dir")]
-    pub report_output_dir: String,
-    /// Optional SIEM webhook URL for alert ingestion.
-    #[serde(default)]
-    pub siem_integration: Option<String>,
-}
-
-fn default_playbooks_dir() -> String {
-    "~/.R.A.I.N./playbooks".into()
-}
-
-fn default_require_approval() -> bool {
-    true
-}
-
-fn default_max_auto_severity() -> String {
-    "low".into()
-}
-
-fn default_report_output_dir() -> String {
-    "~/.R.A.I.N./security-reports".into()
-}
-
-impl Default for SecurityOpsConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            playbooks_dir: default_playbooks_dir(),
-            auto_triage: false,
-            require_approval_for_actions: true,
-            max_auto_severity: default_max_auto_severity(),
-            report_output_dir: default_report_output_dir(),
-            siem_integration: None,
-        }
-    }
-}
-
 // ── Config impl ──────────────────────────────────────────────────
-
-impl Default for Config {
-    fn default() -> Self {
-        let home =
-            UserDirs::new().map_or_else(|| PathBuf::from("."), |u| u.home_dir().to_path_buf());
-        let rain_dir = home.join(".R.A.I.N.");
-
-        Self {
-            workspace_dir: rain_dir.join("workspace"),
-            config_path: rain_dir.join("config.toml"),
-            api_key: None,
-            api_url: None,
-            api_path: None,
-            default_provider: Some("openrouter".to_string()),
-            default_model: Some("anthropic/claude-sonnet-4.6".to_string()),
-            model_providers: HashMap::new(),
-            default_temperature: default_temperature(),
-            provider_timeout_secs: default_provider_timeout_secs(),
-            extra_headers: HashMap::new(),
-            observability: ObservabilityConfig::default(),
-            autonomy: AutonomyConfig::default(),
-            backup: BackupConfig::default(),
-            data_retention: DataRetentionConfig::default(),
-            cloud_ops: CloudOpsConfig::default(),
-            conversational_ai: ConversationalAiConfig::default(),
-            security: SecurityConfig::default(),
-            security_ops: SecurityOpsConfig::default(),
-            runtime: RuntimeConfig::default(),
-            reliability: ReliabilityConfig::default(),
-            scheduler: SchedulerConfig::default(),
-            agent: AgentConfig::default(),
-            pacing: PacingConfig::default(),
-            skills: SkillsConfig::default(),
-            model_routes: Vec::new(),
-            embedding_routes: Vec::new(),
-            heartbeat: HeartbeatConfig::default(),
-            cron: CronConfig::default(),
-            channels_config: ChannelsConfig::default(),
-            memory: MemoryConfig::default(),
-            storage: StorageConfig::default(),
-            tunnel: TunnelConfig::default(),
-            gateway: GatewayConfig::default(),
-            composio: ComposioConfig::default(),
-            microsoft365: Microsoft365Config::default(),
-            secrets: SecretsConfig::default(),
-            browser: BrowserConfig::default(),
-            browser_delegate: crate::tools::browser_delegate::BrowserDelegateConfig::default(),
-            http_request: HttpRequestConfig::default(),
-            multimodal: MultimodalConfig::default(),
-            web_fetch: WebFetchConfig::default(),
-            text_browser: TextBrowserConfig::default(),
-            web_search: WebSearchConfig::default(),
-            project_intel: ProjectIntelConfig::default(),
-            google_workspace: GoogleWorkspaceConfig::default(),
-            proxy: ProxyConfig::default(),
-            identity: IdentityConfig::default(),
-            cost: CostConfig::default(),
-            peripherals: PeripheralsConfig::default(),
-            delegate: DelegateToolConfig::default(),
-            agents: HashMap::new(),
-            swarms: HashMap::new(),
-            hooks: HooksConfig::default(),
-            hardware: HardwareConfig::default(),
-            query_classification: QueryClassificationConfig::default(),
-            transcription: TranscriptionConfig::default(),
-            tts: TtsConfig::default(),
-            mcp: McpConfig::default(),
-            nodes: NodesConfig::default(),
-            workspace: WorkspaceConfig::default(),
-            notion: NotionConfig::default(),
-            jira: JiraConfig::default(),
-            node_transport: NodeTransportConfig::default(),
-            knowledge: KnowledgeConfig::default(),
-            linkedin: LinkedInConfig::default(),
-            plugins: PluginsConfig::default(),
-            locale: None,
-            verifiable_intent: VerifiableIntentConfig::default(),
-            claude_code: ClaudeCodeConfig::default(),
-        }
-    }
-}
 
 fn default_config_and_workspace_dirs() -> Result<(PathBuf, PathBuf)> {
     let config_dir = default_config_dir()?;
@@ -7299,1949 +6466,6 @@ async fn ensure_bootstrap_files(workspace_dir: &Path) -> Result<()> {
     }
 
     Ok(())
-}
-
-impl Config {
-    pub async fn load_or_init() -> Result<Self> {
-        let (default_rain_dir, default_workspace_dir) = default_config_and_workspace_dirs()?;
-
-        let (rain_dir, workspace_dir, resolution_source) =
-            resolve_runtime_config_dirs(&default_rain_dir, &default_workspace_dir).await?;
-
-        let config_path = rain_dir.join("config.toml");
-
-        fs::create_dir_all(&rain_dir)
-            .await
-            .with_context(|| config_dir_creation_error(&rain_dir))?;
-        fs::create_dir_all(&workspace_dir)
-            .await
-            .context("Failed to create workspace directory")?;
-
-        ensure_bootstrap_files(&workspace_dir).await?;
-
-        if config_path.exists() {
-            // Warn if config file is world-readable (may contain API keys)
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-                if let Ok(meta) = fs::metadata(&config_path).await {
-                    if meta.permissions().mode() & 0o004 != 0 {
-                        tracing::warn!(
-                            "Config file {:?} is world-readable (mode {:o}). \
-                             Consider restricting with: chmod 600 {:?}",
-                            config_path,
-                            meta.permissions().mode() & 0o777,
-                            config_path,
-                        );
-                    }
-                }
-            }
-
-            let contents = fs::read_to_string(&config_path)
-                .await
-                .context("Failed to read config file")?;
-
-            // Track ignored/unknown config keys to warn users about silent misconfigurations
-            // (e.g., using [providers.ollama] which doesn't exist instead of top-level api_url)
-            let mut ignored_paths: Vec<String> = Vec::new();
-            let mut config: Config = serde_ignored::deserialize(
-                toml::de::Deserializer::parse(&contents).context("Failed to parse config file")?,
-                |path| {
-                    ignored_paths.push(path.to_string());
-                },
-            )
-            .context("Failed to deserialize config file")?;
-
-            // Warn about each unknown config key.
-            // serde_ignored + #[serde(default)] on nested structs can produce
-            // false positives: parent-level fields get re-reported under the
-            // nested key (e.g. "memory.mem0.auto_hydrate" even though
-            // auto_hydrate belongs to MemoryConfig, not Mem0Config).  We
-            // suppress these by checking whether the leaf key is a known field
-            // on the parent struct.
-            let known_memory_fields: &[&str] = &[
-                "backend",
-                "auto_save",
-                "hygiene_enabled",
-                "archive_after_days",
-                "purge_after_days",
-                "conversation_retention_days",
-                "embedding_provider",
-                "embedding_model",
-                "embedding_dimensions",
-                "vector_weight",
-                "keyword_weight",
-                "min_relevance_score",
-                "embedding_cache_size",
-                "chunk_max_tokens",
-                "response_cache_enabled",
-                "response_cache_ttl_minutes",
-                "response_cache_max_entries",
-                "response_cache_hot_entries",
-                "snapshot_enabled",
-                "snapshot_on_hygiene",
-                "auto_hydrate",
-                "sqlite_open_timeout_secs",
-            ];
-            for path in ignored_paths {
-                // Skip false positives from nested memory sub-sections
-                if path.starts_with("memory.mem0.") || path.starts_with("memory.qdrant.") {
-                    let leaf = path.rsplit('.').next().unwrap_or("");
-                    if known_memory_fields.contains(&leaf) {
-                        continue;
-                    }
-                }
-                tracing::warn!(
-                    "Unknown config key ignored: \"{}\". Check config.toml for typos or deprecated options.",
-                    path
-                );
-            }
-            // Set computed paths that are skipped during serialization
-            config.config_path = config_path.clone();
-            config.workspace_dir = workspace_dir;
-            let store = crate::security::SecretStore::new(&rain_dir, config.secrets.encrypt);
-            decrypt_optional_secret(&store, &mut config.api_key, "config.api_key")?;
-            decrypt_optional_secret(
-                &store,
-                &mut config.composio.api_key,
-                "config.composio.api_key",
-            )?;
-            if let Some(ref mut pinggy) = config.tunnel.pinggy {
-                decrypt_optional_secret(&store, &mut pinggy.token, "config.tunnel.pinggy.token")?;
-            }
-            decrypt_optional_secret(
-                &store,
-                &mut config.microsoft365.client_secret,
-                "config.microsoft365.client_secret",
-            )?;
-
-            decrypt_optional_secret(
-                &store,
-                &mut config.browser.computer_use.api_key,
-                "config.browser.computer_use.api_key",
-            )?;
-
-            decrypt_optional_secret(
-                &store,
-                &mut config.web_search.brave_api_key,
-                "config.web_search.brave_api_key",
-            )?;
-
-            decrypt_optional_secret(
-                &store,
-                &mut config.storage.provider.config.db_url,
-                "config.storage.provider.config.db_url",
-            )?;
-
-            for agent in config.agents.values_mut() {
-                decrypt_optional_secret(&store, &mut agent.api_key, "config.agents.*.api_key")?;
-            }
-
-            // Decrypt TTS provider API keys
-            if let Some(ref mut openai) = config.tts.openai {
-                decrypt_optional_secret(&store, &mut openai.api_key, "config.tts.openai.api_key")?;
-            }
-            if let Some(ref mut elevenlabs) = config.tts.elevenlabs {
-                decrypt_optional_secret(
-                    &store,
-                    &mut elevenlabs.api_key,
-                    "config.tts.elevenlabs.api_key",
-                )?;
-            }
-            if let Some(ref mut google) = config.tts.google {
-                decrypt_optional_secret(&store, &mut google.api_key, "config.tts.google.api_key")?;
-            }
-
-            // Decrypt nested STT provider API keys
-            decrypt_optional_secret(
-                &store,
-                &mut config.transcription.api_key,
-                "config.transcription.api_key",
-            )?;
-            if let Some(ref mut openai) = config.transcription.openai {
-                decrypt_optional_secret(
-                    &store,
-                    &mut openai.api_key,
-                    "config.transcription.openai.api_key",
-                )?;
-            }
-            if let Some(ref mut deepgram) = config.transcription.deepgram {
-                decrypt_optional_secret(
-                    &store,
-                    &mut deepgram.api_key,
-                    "config.transcription.deepgram.api_key",
-                )?;
-            }
-            if let Some(ref mut assemblyai) = config.transcription.assemblyai {
-                decrypt_optional_secret(
-                    &store,
-                    &mut assemblyai.api_key,
-                    "config.transcription.assemblyai.api_key",
-                )?;
-            }
-            if let Some(ref mut google) = config.transcription.google {
-                decrypt_optional_secret(
-                    &store,
-                    &mut google.api_key,
-                    "config.transcription.google.api_key",
-                )?;
-            }
-
-            #[cfg(feature = "channel-nostr")]
-            if let Some(ref mut ns) = config.channels_config.nostr {
-                decrypt_secret(
-                    &store,
-                    &mut ns.private_key,
-                    "config.channels_config.nostr.private_key",
-                )?;
-            }
-            if let Some(ref mut fs) = config.channels_config.feishu {
-                decrypt_secret(
-                    &store,
-                    &mut fs.app_secret,
-                    "config.channels_config.feishu.app_secret",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut fs.encrypt_key,
-                    "config.channels_config.feishu.encrypt_key",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut fs.verification_token,
-                    "config.channels_config.feishu.verification_token",
-                )?;
-            }
-
-            // Decrypt channel secrets
-            if let Some(ref mut tg) = config.channels_config.telegram {
-                decrypt_secret(
-                    &store,
-                    &mut tg.bot_token,
-                    "config.channels_config.telegram.bot_token",
-                )?;
-            }
-            if let Some(ref mut dc) = config.channels_config.discord {
-                decrypt_secret(
-                    &store,
-                    &mut dc.bot_token,
-                    "config.channels_config.discord.bot_token",
-                )?;
-            }
-            if let Some(ref mut sl) = config.channels_config.slack {
-                decrypt_secret(
-                    &store,
-                    &mut sl.bot_token,
-                    "config.channels_config.slack.bot_token",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut sl.app_token,
-                    "config.channels_config.slack.app_token",
-                )?;
-            }
-            if let Some(ref mut mm) = config.channels_config.mattermost {
-                decrypt_secret(
-                    &store,
-                    &mut mm.bot_token,
-                    "config.channels_config.mattermost.bot_token",
-                )?;
-            }
-            if let Some(ref mut mx) = config.channels_config.matrix {
-                decrypt_secret(
-                    &store,
-                    &mut mx.access_token,
-                    "config.channels_config.matrix.access_token",
-                )?;
-            }
-            if let Some(ref mut wa) = config.channels_config.whatsapp {
-                decrypt_optional_secret(
-                    &store,
-                    &mut wa.access_token,
-                    "config.channels_config.whatsapp.access_token",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut wa.app_secret,
-                    "config.channels_config.whatsapp.app_secret",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut wa.verify_token,
-                    "config.channels_config.whatsapp.verify_token",
-                )?;
-            }
-            if let Some(ref mut lq) = config.channels_config.linq {
-                decrypt_secret(
-                    &store,
-                    &mut lq.api_token,
-                    "config.channels_config.linq.api_token",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut lq.signing_secret,
-                    "config.channels_config.linq.signing_secret",
-                )?;
-            }
-            if let Some(ref mut wt) = config.channels_config.wati {
-                decrypt_secret(
-                    &store,
-                    &mut wt.api_token,
-                    "config.channels_config.wati.api_token",
-                )?;
-            }
-            if let Some(ref mut nc) = config.channels_config.nextcloud_talk {
-                decrypt_secret(
-                    &store,
-                    &mut nc.app_token,
-                    "config.channels_config.nextcloud_talk.app_token",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut nc.webhook_secret,
-                    "config.channels_config.nextcloud_talk.webhook_secret",
-                )?;
-            }
-            if let Some(ref mut em) = config.channels_config.email {
-                decrypt_secret(
-                    &store,
-                    &mut em.password,
-                    "config.channels_config.email.password",
-                )?;
-            }
-            if let Some(ref mut irc) = config.channels_config.irc {
-                decrypt_optional_secret(
-                    &store,
-                    &mut irc.server_password,
-                    "config.channels_config.irc.server_password",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut irc.nickserv_password,
-                    "config.channels_config.irc.nickserv_password",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut irc.sasl_password,
-                    "config.channels_config.irc.sasl_password",
-                )?;
-            }
-            if let Some(ref mut lk) = config.channels_config.lark {
-                decrypt_secret(
-                    &store,
-                    &mut lk.app_secret,
-                    "config.channels_config.lark.app_secret",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut lk.encrypt_key,
-                    "config.channels_config.lark.encrypt_key",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut lk.verification_token,
-                    "config.channels_config.lark.verification_token",
-                )?;
-            }
-            if let Some(ref mut fs) = config.channels_config.feishu {
-                decrypt_secret(
-                    &store,
-                    &mut fs.app_secret,
-                    "config.channels_config.feishu.app_secret",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut fs.encrypt_key,
-                    "config.channels_config.feishu.encrypt_key",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut fs.verification_token,
-                    "config.channels_config.feishu.verification_token",
-                )?;
-            }
-            if let Some(ref mut dt) = config.channels_config.dingtalk {
-                decrypt_secret(
-                    &store,
-                    &mut dt.client_secret,
-                    "config.channels_config.dingtalk.client_secret",
-                )?;
-            }
-            if let Some(ref mut wc) = config.channels_config.wecom {
-                decrypt_secret(
-                    &store,
-                    &mut wc.webhook_key,
-                    "config.channels_config.wecom.webhook_key",
-                )?;
-            }
-            if let Some(ref mut qq) = config.channels_config.qq {
-                decrypt_secret(
-                    &store,
-                    &mut qq.app_secret,
-                    "config.channels_config.qq.app_secret",
-                )?;
-            }
-            if let Some(ref mut wh) = config.channels_config.webhook {
-                decrypt_optional_secret(
-                    &store,
-                    &mut wh.secret,
-                    "config.channels_config.webhook.secret",
-                )?;
-            }
-            if let Some(ref mut ct) = config.channels_config.clawdtalk {
-                decrypt_secret(
-                    &store,
-                    &mut ct.api_key,
-                    "config.channels_config.clawdtalk.api_key",
-                )?;
-                decrypt_optional_secret(
-                    &store,
-                    &mut ct.webhook_secret,
-                    "config.channels_config.clawdtalk.webhook_secret",
-                )?;
-            }
-
-            // Decrypt gateway paired tokens
-            for token in &mut config.gateway.paired_tokens {
-                decrypt_secret(&store, token, "config.gateway.paired_tokens[]")?;
-            }
-
-            // Decrypt Nevis IAM secret
-            decrypt_optional_secret(
-                &store,
-                &mut config.security.nevis.client_secret,
-                "config.security.nevis.client_secret",
-            )?;
-
-            // Notion API key (top-level, not in ChannelsConfig)
-            if !config.notion.api_key.is_empty() {
-                decrypt_secret(&store, &mut config.notion.api_key, "config.notion.api_key")?;
-            }
-
-            // Jira API token
-            if !config.jira.api_token.is_empty() {
-                decrypt_secret(&store, &mut config.jira.api_token, "config.jira.api_token")?;
-            }
-
-            config.apply_env_overrides();
-            config.validate()?;
-            tracing::info!(
-                path = %config.config_path.display(),
-                workspace = %config.workspace_dir.display(),
-                source = resolution_source.as_str(),
-                initialized = true,
-                "Config loaded"
-            );
-            Ok(config)
-        } else {
-            let mut config = Config::default();
-            config.config_path = config_path.clone();
-            config.workspace_dir = workspace_dir;
-            config.save().await?;
-
-            // Restrict permissions on newly created config file (may contain API keys)
-            #[cfg(unix)]
-            {
-                use std::{fs::Permissions, os::unix::fs::PermissionsExt};
-                let _ = fs::set_permissions(&config_path, Permissions::from_mode(0o600)).await;
-            }
-
-            config.apply_env_overrides();
-            config.validate()?;
-            tracing::info!(
-                path = %config.config_path.display(),
-                workspace = %config.workspace_dir.display(),
-                source = resolution_source.as_str(),
-                initialized = true,
-                "Config loaded"
-            );
-            Ok(config)
-        }
-    }
-
-    fn lookup_model_provider_profile(
-        &self,
-        provider_name: &str,
-    ) -> Option<(String, ModelProviderConfig)> {
-        let needle = provider_name.trim();
-        if needle.is_empty() {
-            return None;
-        }
-
-        self.model_providers
-            .iter()
-            .find(|(name, _)| name.eq_ignore_ascii_case(needle))
-            .map(|(name, profile)| (name.clone(), profile.clone()))
-    }
-
-    fn apply_named_model_provider_profile(&mut self) {
-        let Some(current_provider) = self.default_provider.clone() else {
-            return;
-        };
-
-        let Some((profile_key, profile)) = self.lookup_model_provider_profile(&current_provider)
-        else {
-            return;
-        };
-
-        let base_url = profile
-            .base_url
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(ToString::to_string);
-
-        if self
-            .api_url
-            .as_deref()
-            .map(str::trim)
-            .is_none_or(|value| value.is_empty())
-        {
-            if let Some(base_url) = base_url.as_ref() {
-                self.api_url = Some(base_url.clone());
-            }
-        }
-
-        // Propagate api_path from the profile when not already set at top level.
-        if self.api_path.is_none() {
-            if let Some(ref path) = profile.api_path {
-                let trimmed = path.trim();
-                if !trimmed.is_empty() {
-                    self.api_path = Some(trimmed.to_string());
-                }
-            }
-        }
-
-        if profile.requires_openai_auth
-            && self
-                .api_key
-                .as_deref()
-                .map(str::trim)
-                .is_none_or(|value| value.is_empty())
-        {
-            let codex_key = std::env::var("OPENAI_API_KEY")
-                .ok()
-                .map(|value| value.trim().to_string())
-                .filter(|value| !value.is_empty())
-                .or_else(read_codex_openai_api_key);
-            if let Some(codex_key) = codex_key {
-                self.api_key = Some(codex_key);
-            }
-        }
-
-        let normalized_wire_api = profile.wire_api.as_deref().and_then(normalize_wire_api);
-        let profile_name = profile
-            .name
-            .as_deref()
-            .map(str::trim)
-            .filter(|value| !value.is_empty());
-
-        if normalized_wire_api == Some("responses") {
-            self.default_provider = Some("openai-codex".to_string());
-            return;
-        }
-
-        if let Some(profile_name) = profile_name {
-            if !profile_name.eq_ignore_ascii_case(&profile_key) {
-                self.default_provider = Some(profile_name.to_string());
-                return;
-            }
-        }
-
-        if let Some(base_url) = base_url {
-            self.default_provider = Some(format!("custom:{base_url}"));
-        }
-    }
-
-    /// Validate configuration values that would cause runtime failures.
-    ///
-    /// Called after TOML deserialization and env-override application to catch
-    /// obviously invalid values early instead of failing at arbitrary runtime points.
-    pub fn validate(&self) -> Result<()> {
-        // Tunnel — OpenVPN
-        if self.tunnel.provider.trim() == "openvpn" {
-            let openvpn = self.tunnel.openvpn.as_ref().ok_or_else(|| {
-                anyhow::anyhow!("tunnel.provider='openvpn' requires [tunnel.openvpn]")
-            })?;
-
-            if openvpn.config_file.trim().is_empty() {
-                anyhow::bail!("tunnel.openvpn.config_file must not be empty");
-            }
-            if openvpn.connect_timeout_secs == 0 {
-                anyhow::bail!("tunnel.openvpn.connect_timeout_secs must be greater than 0");
-            }
-        }
-
-        // Gateway
-        if self.gateway.host.trim().is_empty() {
-            anyhow::bail!("gateway.host must not be empty");
-        }
-        if let Some(ref prefix) = self.gateway.path_prefix {
-            // Validate the raw value — no silent trimming so the stored
-            // value is exactly what was validated.
-            if !prefix.is_empty() {
-                if !prefix.starts_with('/') {
-                    anyhow::bail!("gateway.path_prefix must start with '/'");
-                }
-                if prefix.ends_with('/') {
-                    anyhow::bail!("gateway.path_prefix must not end with '/' (including bare '/')");
-                }
-                // Reject characters unsafe for URL paths or HTML/JS injection.
-                // Whitespace is intentionally excluded from the allowed set.
-                if let Some(bad) = prefix.chars().find(|c| {
-                    !matches!(c, '/' | '-' | '_' | '.' | '~'
-                        | 'a'..='z' | 'A'..='Z' | '0'..='9'
-                        | '!' | '$' | '&' | '\'' | '(' | ')' | '*' | '+' | ',' | ';' | '='
-                        | ':' | '@')
-                }) {
-                    anyhow::bail!(
-                        "gateway.path_prefix contains invalid character '{bad}'; \
-                         only unreserved and sub-delim URI characters are allowed"
-                    );
-                }
-            }
-        }
-
-        // Autonomy
-        if self.autonomy.max_actions_per_hour == 0 {
-            anyhow::bail!("autonomy.max_actions_per_hour must be greater than 0");
-        }
-        for (i, env_name) in self.autonomy.shell_env_passthrough.iter().enumerate() {
-            if !is_valid_env_var_name(env_name) {
-                anyhow::bail!(
-                    "autonomy.shell_env_passthrough[{i}] is invalid ({env_name}); expected [A-Za-z_][A-Za-z0-9_]*"
-                );
-            }
-        }
-
-        // Security OTP / estop
-        if self.security.otp.challenge_max_attempts == 0 {
-            anyhow::bail!("security.otp.challenge_max_attempts must be greater than 0");
-        }
-        if self.security.otp.token_ttl_secs == 0 {
-            anyhow::bail!("security.otp.token_ttl_secs must be greater than 0");
-        }
-        if self.security.otp.cache_valid_secs == 0 {
-            anyhow::bail!("security.otp.cache_valid_secs must be greater than 0");
-        }
-        if self.security.otp.cache_valid_secs < self.security.otp.token_ttl_secs {
-            anyhow::bail!(
-                "security.otp.cache_valid_secs must be greater than or equal to security.otp.token_ttl_secs"
-            );
-        }
-        if self.security.otp.challenge_max_attempts == 0 {
-            anyhow::bail!("security.otp.challenge_max_attempts must be greater than 0");
-        }
-        for (i, action) in self.security.otp.gated_actions.iter().enumerate() {
-            let normalized = action.trim();
-            if normalized.is_empty() {
-                anyhow::bail!("security.otp.gated_actions[{i}] must not be empty");
-            }
-            if !normalized
-                .chars()
-                .all(|c| c.is_ascii_alphanumeric() || c == '_' || c == '-')
-            {
-                anyhow::bail!(
-                    "security.otp.gated_actions[{i}] contains invalid characters: {normalized}"
-                );
-            }
-        }
-        DomainMatcher::new(
-            &self.security.otp.gated_domains,
-            &self.security.otp.gated_domain_categories,
-        )
-        .with_context(|| {
-            "Invalid security.otp.gated_domains or security.otp.gated_domain_categories"
-        })?;
-        if self.security.estop.state_file.trim().is_empty() {
-            anyhow::bail!("security.estop.state_file must not be empty");
-        }
-
-        // Scheduler
-        if self.scheduler.max_concurrent == 0 {
-            anyhow::bail!("scheduler.max_concurrent must be greater than 0");
-        }
-        if self.scheduler.max_tasks == 0 {
-            anyhow::bail!("scheduler.max_tasks must be greater than 0");
-        }
-
-        // Model routes
-        for (i, route) in self.model_routes.iter().enumerate() {
-            if route.hint.trim().is_empty() {
-                anyhow::bail!("model_routes[{i}].hint must not be empty");
-            }
-            if route.provider.trim().is_empty() {
-                anyhow::bail!("model_routes[{i}].provider must not be empty");
-            }
-            if route.model.trim().is_empty() {
-                anyhow::bail!("model_routes[{i}].model must not be empty");
-            }
-        }
-
-        // Embedding routes
-        for (i, route) in self.embedding_routes.iter().enumerate() {
-            if route.hint.trim().is_empty() {
-                anyhow::bail!("embedding_routes[{i}].hint must not be empty");
-            }
-            if route.provider.trim().is_empty() {
-                anyhow::bail!("embedding_routes[{i}].provider must not be empty");
-            }
-            if route.model.trim().is_empty() {
-                anyhow::bail!("embedding_routes[{i}].model must not be empty");
-            }
-        }
-
-        for (profile_key, profile) in &self.model_providers {
-            let profile_name = profile_key.trim();
-            if profile_name.is_empty() {
-                anyhow::bail!("model_providers contains an empty profile name");
-            }
-
-            let has_name = profile
-                .name
-                .as_deref()
-                .map(str::trim)
-                .is_some_and(|value| !value.is_empty());
-            let has_base_url = profile
-                .base_url
-                .as_deref()
-                .map(str::trim)
-                .is_some_and(|value| !value.is_empty());
-
-            if !has_name && !has_base_url {
-                anyhow::bail!(
-                    "model_providers.{profile_name} must define at least one of `name` or `base_url`"
-                );
-            }
-
-            if let Some(base_url) = profile.base_url.as_deref().map(str::trim) {
-                if !base_url.is_empty() {
-                    let parsed = reqwest::Url::parse(base_url).with_context(|| {
-                        format!("model_providers.{profile_name}.base_url is not a valid URL")
-                    })?;
-                    if !matches!(parsed.scheme(), "http" | "https") {
-                        anyhow::bail!(
-                            "model_providers.{profile_name}.base_url must use http/https"
-                        );
-                    }
-                }
-            }
-
-            if let Some(wire_api) = profile.wire_api.as_deref().map(str::trim) {
-                if !wire_api.is_empty() && normalize_wire_api(wire_api).is_none() {
-                    anyhow::bail!(
-                        "model_providers.{profile_name}.wire_api must be one of: responses, chat_completions"
-                    );
-                }
-            }
-        }
-
-        // Ollama cloud-routing safety checks
-        if self
-            .default_provider
-            .as_deref()
-            .is_some_and(|provider| provider.trim().eq_ignore_ascii_case("ollama"))
-            && self
-                .default_model
-                .as_deref()
-                .is_some_and(|model| model.trim().ends_with(":cloud"))
-        {
-            if is_local_ollama_endpoint(self.api_url.as_deref()) {
-                anyhow::bail!(
-                    "default_model uses ':cloud' with provider 'ollama', but api_url is local or unset. Set api_url to a remote Ollama endpoint (for example https://ollama.com)."
-                );
-            }
-
-            if !has_ollama_cloud_credential(self.api_key.as_deref()) {
-                anyhow::bail!(
-                    "default_model uses ':cloud' with provider 'ollama', but no API key is configured. Set api_key or OLLAMA_API_KEY."
-                );
-            }
-        }
-
-        // Microsoft 365
-        if self.microsoft365.enabled {
-            let tenant = self
-                .microsoft365
-                .tenant_id
-                .as_deref()
-                .map(str::trim)
-                .filter(|s| !s.is_empty());
-            if tenant.is_none() {
-                anyhow::bail!(
-                    "microsoft365.tenant_id must not be empty when microsoft365 is enabled"
-                );
-            }
-            let client = self
-                .microsoft365
-                .client_id
-                .as_deref()
-                .map(str::trim)
-                .filter(|s| !s.is_empty());
-            if client.is_none() {
-                anyhow::bail!(
-                    "microsoft365.client_id must not be empty when microsoft365 is enabled"
-                );
-            }
-            let flow = self.microsoft365.auth_flow.trim();
-            if flow != "client_credentials" && flow != "device_code" {
-                anyhow::bail!(
-                    "microsoft365.auth_flow must be 'client_credentials' or 'device_code'"
-                );
-            }
-            if flow == "client_credentials"
-                && self
-                    .microsoft365
-                    .client_secret
-                    .as_deref()
-                    .map_or(true, |s| s.trim().is_empty())
-            {
-                anyhow::bail!(
-                    "microsoft365.client_secret must not be empty when auth_flow is 'client_credentials'"
-                );
-            }
-        }
-
-        // Microsoft 365
-        if self.microsoft365.enabled {
-            let tenant = self
-                .microsoft365
-                .tenant_id
-                .as_deref()
-                .map(str::trim)
-                .filter(|s| !s.is_empty());
-            if tenant.is_none() {
-                anyhow::bail!(
-                    "microsoft365.tenant_id must not be empty when microsoft365 is enabled"
-                );
-            }
-            let client = self
-                .microsoft365
-                .client_id
-                .as_deref()
-                .map(str::trim)
-                .filter(|s| !s.is_empty());
-            if client.is_none() {
-                anyhow::bail!(
-                    "microsoft365.client_id must not be empty when microsoft365 is enabled"
-                );
-            }
-            let flow = self.microsoft365.auth_flow.trim();
-            if flow != "client_credentials" && flow != "device_code" {
-                anyhow::bail!("microsoft365.auth_flow must be client_credentials or device_code");
-            }
-            if flow == "client_credentials"
-                && self
-                    .microsoft365
-                    .client_secret
-                    .as_deref()
-                    .map_or(true, |s| s.trim().is_empty())
-            {
-                anyhow::bail!("microsoft365.client_secret must not be empty when auth_flow is client_credentials");
-            }
-        }
-
-        // MCP
-        if self.mcp.enabled {
-            validate_mcp_config(&self.mcp)?;
-        }
-
-        // Knowledge graph
-        if self.knowledge.enabled {
-            if self.knowledge.max_nodes == 0 {
-                anyhow::bail!("knowledge.max_nodes must be greater than 0");
-            }
-            if self.knowledge.db_path.trim().is_empty() {
-                anyhow::bail!("knowledge.db_path must not be empty");
-            }
-        }
-
-        // Google Workspace allowed_services validation
-        let mut seen_gws_services = std::collections::HashSet::new();
-        for (i, service) in self.google_workspace.allowed_services.iter().enumerate() {
-            let normalized = service.trim();
-            if normalized.is_empty() {
-                anyhow::bail!("google_workspace.allowed_services[{i}] must not be empty");
-            }
-            if !normalized
-                .chars()
-                .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '_' || c == '-')
-            {
-                anyhow::bail!(
-                    "google_workspace.allowed_services[{i}] contains invalid characters: {normalized}"
-                );
-            }
-            if !seen_gws_services.insert(normalized.to_string()) {
-                anyhow::bail!(
-                    "google_workspace.allowed_services contains duplicate entry: {normalized}"
-                );
-            }
-        }
-
-        // Build the effective allowed-services set for cross-validation.
-        // When the operator leaves allowed_services empty the tool falls back to
-        // DEFAULT_GWS_SERVICES; use the same constant here so validation is
-        // consistent in both cases.
-        let effective_services: std::collections::HashSet<&str> =
-            if self.google_workspace.allowed_services.is_empty() {
-                DEFAULT_GWS_SERVICES.iter().copied().collect()
-            } else {
-                self.google_workspace
-                    .allowed_services
-                    .iter()
-                    .map(|s| s.trim())
-                    .collect()
-            };
-
-        let mut seen_gws_operations = std::collections::HashSet::new();
-        for (i, operation) in self.google_workspace.allowed_operations.iter().enumerate() {
-            let service = operation.service.trim();
-            let resource = operation.resource.trim();
-
-            if service.is_empty() {
-                anyhow::bail!("google_workspace.allowed_operations[{i}].service must not be empty");
-            }
-            if resource.is_empty() {
-                anyhow::bail!(
-                    "google_workspace.allowed_operations[{i}].resource must not be empty"
-                );
-            }
-
-            if !effective_services.contains(service) {
-                anyhow::bail!(
-                    "google_workspace.allowed_operations[{i}].service '{service}' is not in the \
-                     effective allowed_services; this entry can never match at runtime"
-                );
-            }
-            if !service
-                .chars()
-                .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '_' || c == '-')
-            {
-                anyhow::bail!(
-                    "google_workspace.allowed_operations[{i}].service contains invalid characters: {service}"
-                );
-            }
-            if !resource
-                .chars()
-                .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '_' || c == '-')
-            {
-                anyhow::bail!(
-                    "google_workspace.allowed_operations[{i}].resource contains invalid characters: {resource}"
-                );
-            }
-
-            if let Some(ref sub_resource) = operation.sub_resource {
-                let sub = sub_resource.trim();
-                if sub.is_empty() {
-                    anyhow::bail!(
-                        "google_workspace.allowed_operations[{i}].sub_resource must not be empty when present"
-                    );
-                }
-                if !sub
-                    .chars()
-                    .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '_' || c == '-')
-                {
-                    anyhow::bail!(
-                        "google_workspace.allowed_operations[{i}].sub_resource contains invalid characters: {sub}"
-                    );
-                }
-            }
-
-            if operation.methods.is_empty() {
-                anyhow::bail!("google_workspace.allowed_operations[{i}].methods must not be empty");
-            }
-
-            let mut seen_methods = std::collections::HashSet::new();
-            for (j, method) in operation.methods.iter().enumerate() {
-                let normalized = method.trim();
-                if normalized.is_empty() {
-                    anyhow::bail!(
-                        "google_workspace.allowed_operations[{i}].methods[{j}] must not be empty"
-                    );
-                }
-                if !normalized
-                    .chars()
-                    .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '_' || c == '-')
-                {
-                    anyhow::bail!(
-                        "google_workspace.allowed_operations[{i}].methods[{j}] contains invalid characters: {normalized}"
-                    );
-                }
-                if !seen_methods.insert(normalized.to_string()) {
-                    anyhow::bail!(
-                        "google_workspace.allowed_operations[{i}].methods contains duplicate entry: {normalized}"
-                    );
-                }
-            }
-
-            let sub_key = operation
-                .sub_resource
-                .as_deref()
-                .map(str::trim)
-                .unwrap_or("");
-            let operation_key = format!("{service}:{resource}:{sub_key}");
-            if !seen_gws_operations.insert(operation_key.clone()) {
-                anyhow::bail!(
-                    "google_workspace.allowed_operations contains duplicate service/resource/sub_resource entry: {operation_key}"
-                );
-            }
-        }
-
-        // Project intelligence
-        if self.project_intel.enabled {
-            let lang = &self.project_intel.default_language;
-            if !["en", "de", "fr", "it"].contains(&lang.as_str()) {
-                anyhow::bail!(
-                    "project_intel.default_language must be one of: en, de, fr, it (got '{lang}')"
-                );
-            }
-            let sens = &self.project_intel.risk_sensitivity;
-            if !["low", "medium", "high"].contains(&sens.as_str()) {
-                anyhow::bail!(
-                    "project_intel.risk_sensitivity must be one of: low, medium, high (got '{sens}')"
-                );
-            }
-            if let Some(ref tpl_dir) = self.project_intel.templates_dir {
-                let path = std::path::Path::new(tpl_dir);
-                if !path.exists() {
-                    anyhow::bail!("project_intel.templates_dir path does not exist: {tpl_dir}");
-                }
-            }
-        }
-
-        // Proxy (delegate to existing validation)
-        self.proxy.validate()?;
-        self.cloud_ops.validate()?;
-
-        // Notion
-        if self.notion.enabled {
-            if self.notion.database_id.trim().is_empty() {
-                anyhow::bail!("notion.database_id must not be empty when notion.enabled = true");
-            }
-            if self.notion.poll_interval_secs == 0 {
-                anyhow::bail!("notion.poll_interval_secs must be greater than 0");
-            }
-            if self.notion.max_concurrent == 0 {
-                anyhow::bail!("notion.max_concurrent must be greater than 0");
-            }
-            if self.notion.status_property.trim().is_empty() {
-                anyhow::bail!("notion.status_property must not be empty");
-            }
-            if self.notion.input_property.trim().is_empty() {
-                anyhow::bail!("notion.input_property must not be empty");
-            }
-            if self.notion.result_property.trim().is_empty() {
-                anyhow::bail!("notion.result_property must not be empty");
-            }
-        }
-
-        // Pinggy tunnel region — validate allowed values (case-insensitive, auto-lowercased at runtime).
-        if let Some(ref pinggy) = self.tunnel.pinggy {
-            if let Some(ref region) = pinggy.region {
-                let r = region.trim().to_ascii_lowercase();
-                if !r.is_empty() && !matches!(r.as_str(), "us" | "eu" | "ap" | "br" | "au") {
-                    anyhow::bail!(
-                        "tunnel.pinggy.region must be one of: us, eu, ap, br, au (or omitted for auto)"
-                    );
-                }
-            }
-        }
-
-        // Jira
-        if self.jira.enabled {
-            if self.jira.base_url.trim().is_empty() {
-                anyhow::bail!("jira.base_url must not be empty when jira.enabled = true");
-            }
-            if self.jira.email.trim().is_empty() {
-                anyhow::bail!("jira.email must not be empty when jira.enabled = true");
-            }
-            if self.jira.api_token.trim().is_empty()
-                && std::env::var("JIRA_API_TOKEN")
-                    .unwrap_or_default()
-                    .trim()
-                    .is_empty()
-            {
-                anyhow::bail!(
-                    "jira.api_token must be set (or JIRA_API_TOKEN env var) when jira.enabled = true"
-                );
-            }
-            let valid_actions = ["get_ticket", "search_tickets", "comment_ticket"];
-            for action in &self.jira.allowed_actions {
-                if !valid_actions.contains(&action.as_str()) {
-                    anyhow::bail!(
-                        "jira.allowed_actions contains unknown action: '{}'. \
-                         Valid: get_ticket, search_tickets, comment_ticket",
-                        action
-                    );
-                }
-            }
-        }
-
-        // Nevis IAM — delegate to NevisConfig::validate() for field-level checks
-        if let Err(msg) = self.security.nevis.validate() {
-            anyhow::bail!("security.nevis: {msg}");
-        }
-
-        // Delegate agent timeouts
-        const MAX_DELEGATE_TIMEOUT_SECS: u64 = 3600;
-        for (name, agent) in &self.agents {
-            if let Some(timeout) = agent.timeout_secs {
-                if timeout == 0 {
-                    anyhow::bail!("agents.{name}.timeout_secs must be greater than 0");
-                }
-                if timeout > MAX_DELEGATE_TIMEOUT_SECS {
-                    anyhow::bail!(
-                        "agents.{name}.timeout_secs exceeds max {MAX_DELEGATE_TIMEOUT_SECS}"
-                    );
-                }
-            }
-            if let Some(timeout) = agent.agentic_timeout_secs {
-                if timeout == 0 {
-                    anyhow::bail!("agents.{name}.agentic_timeout_secs must be greater than 0");
-                }
-                if timeout > MAX_DELEGATE_TIMEOUT_SECS {
-                    anyhow::bail!(
-                        "agents.{name}.agentic_timeout_secs exceeds max {MAX_DELEGATE_TIMEOUT_SECS}"
-                    );
-                }
-            }
-        }
-
-        // Transcription
-        {
-            let dp = self.transcription.default_provider.trim();
-            match dp {
-                "groq" | "openai" | "deepgram" | "assemblyai" | "google" | "local_whisper" => {}
-                other => {
-                    anyhow::bail!(
-                        "transcription.default_provider must be one of: groq, openai, deepgram, assemblyai, google, local_whisper (got '{other}')"
-                    );
-                }
-            }
-        }
-
-        // Delegate tool global defaults
-        if self.delegate.timeout_secs == 0 {
-            anyhow::bail!("delegate.timeout_secs must be greater than 0");
-        }
-        if self.delegate.agentic_timeout_secs == 0 {
-            anyhow::bail!("delegate.agentic_timeout_secs must be greater than 0");
-        }
-
-        // Per-agent delegate timeout overrides
-        for (name, agent) in &self.agents {
-            if let Some(t) = agent.timeout_secs {
-                if t == 0 {
-                    anyhow::bail!("agents.{name}.timeout_secs must be greater than 0");
-                }
-            }
-            if let Some(t) = agent.agentic_timeout_secs {
-                if t == 0 {
-                    anyhow::bail!("agents.{name}.agentic_timeout_secs must be greater than 0");
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Apply environment variable overrides to config
-    pub fn apply_env_overrides(&mut self) {
-        // API Key: rain_API_KEY or API_KEY (generic)
-        if let Ok(key) = std::env::var("rain_API_KEY").or_else(|_| std::env::var("API_KEY")) {
-            if !key.is_empty() {
-                self.api_key = Some(key);
-            }
-        }
-        // API Key: GLM_API_KEY overrides when provider is a GLM/Zhipu variant.
-        if self.default_provider.as_deref().is_some_and(is_glm_alias) {
-            if let Ok(key) = std::env::var("GLM_API_KEY") {
-                if !key.is_empty() {
-                    self.api_key = Some(key);
-                }
-            }
-        }
-
-        // API Key: ZAI_API_KEY overrides when provider is a Z.AI variant.
-        if self.default_provider.as_deref().is_some_and(is_zai_alias) {
-            if let Ok(key) = std::env::var("ZAI_API_KEY") {
-                if !key.is_empty() {
-                    self.api_key = Some(key);
-                }
-            }
-        }
-
-        // Provider override precedence:
-        // 1) rain_PROVIDER always wins when set.
-        // 2) rain_MODEL_PROVIDER/MODEL_PROVIDER (Codex app-server style).
-        // 3) Legacy PROVIDER is honored only when config still uses default provider.
-        if let Ok(provider) = std::env::var("rain_PROVIDER") {
-            if !provider.is_empty() {
-                self.default_provider = Some(provider);
-            }
-        } else if let Ok(provider) =
-            std::env::var("rain_MODEL_PROVIDER").or_else(|_| std::env::var("MODEL_PROVIDER"))
-        {
-            if !provider.is_empty() {
-                self.default_provider = Some(provider);
-            }
-        } else if let Ok(provider) = std::env::var("PROVIDER") {
-            let should_apply_legacy_provider =
-                self.default_provider.as_deref().map_or(true, |configured| {
-                    configured.trim().eq_ignore_ascii_case("openrouter")
-                });
-            if should_apply_legacy_provider && !provider.is_empty() {
-                self.default_provider = Some(provider);
-            }
-        }
-
-        // Model: rain_MODEL or MODEL
-        if let Ok(model) = std::env::var("rain_MODEL").or_else(|_| std::env::var("MODEL")) {
-            if !model.is_empty() {
-                self.default_model = Some(model);
-            }
-        }
-
-        // Provider HTTP timeout: rain_PROVIDER_TIMEOUT_SECS
-        if let Ok(timeout_secs) = std::env::var("rain_PROVIDER_TIMEOUT_SECS") {
-            if let Ok(timeout_secs) = timeout_secs.parse::<u64>() {
-                if timeout_secs > 0 {
-                    self.provider_timeout_secs = timeout_secs;
-                }
-            }
-        }
-
-        // Extra provider headers: rain_EXTRA_HEADERS
-        // Format: "Key:Value,Key2:Value2"
-        // Env var headers override config file headers with the same name.
-        if let Ok(raw) = std::env::var("rain_EXTRA_HEADERS") {
-            for header in parse_extra_headers_env(&raw) {
-                self.extra_headers.insert(header.0, header.1);
-            }
-        }
-
-        // Apply named provider profile remapping (Codex app-server compatibility).
-        self.apply_named_model_provider_profile();
-
-        // Workspace directory: rain_WORKSPACE
-        if let Ok(workspace) = std::env::var("rain_WORKSPACE") {
-            if !workspace.is_empty() {
-                let expanded = expand_tilde_path(&workspace);
-                let (_, workspace_dir) = resolve_config_dir_for_workspace(&expanded);
-                self.workspace_dir = workspace_dir;
-            }
-        }
-
-        // Open-skills opt-in flag: rain_OPEN_SKILLS_ENABLED
-        if let Ok(flag) = std::env::var("rain_OPEN_SKILLS_ENABLED") {
-            if !flag.trim().is_empty() {
-                match flag.trim().to_ascii_lowercase().as_str() {
-                    "1" | "true" | "yes" | "on" => self.skills.open_skills_enabled = true,
-                    "0" | "false" | "no" | "off" => self.skills.open_skills_enabled = false,
-                    _ => tracing::warn!(
-                        "Ignoring invalid rain_OPEN_SKILLS_ENABLED (valid: 1|0|true|false|yes|no|on|off)"
-                    ),
-                }
-            }
-        }
-
-        // Open-skills directory override: rain_OPEN_SKILLS_DIR
-        if let Ok(path) = std::env::var("rain_OPEN_SKILLS_DIR") {
-            let trimmed = path.trim();
-            if !trimmed.is_empty() {
-                self.skills.open_skills_dir = Some(trimmed.to_string());
-            }
-        }
-
-        // Skills script-file audit override: rain_SKILLS_ALLOW_SCRIPTS
-        if let Ok(flag) = std::env::var("rain_SKILLS_ALLOW_SCRIPTS") {
-            if !flag.trim().is_empty() {
-                match flag.trim().to_ascii_lowercase().as_str(){
-                    "1" | "true" | "yes" | "on" => self.skills.allow_scripts = true,
-                    "0" | "false" | "no" | "off" => self.skills.allow_scripts = false,
-                    _ => tracing::warn!(
-                        "Ignoring invalid rain_SKILLS_ALLOW_SCRIPTS (valid: 1|0|true|false|yes|no|on|off)"
-                    ),
-                }
-            }
-        }
-
-        // Skills prompt mode override: rain_SKILLS_PROMPT_MODE
-        if let Ok(mode) = std::env::var("rain_SKILLS_PROMPT_MODE") {
-            if !mode.trim().is_empty() {
-                if let Some(parsed) = parse_skills_prompt_injection_mode(&mode) {
-                    self.skills.prompt_injection_mode = parsed;
-                } else {
-                    tracing::warn!(
-                        "Ignoring invalid rain_SKILLS_PROMPT_MODE (valid: full|compact)"
-                    );
-                }
-            }
-        }
-
-        // Gateway port: rain_GATEWAY_PORT or PORT
-        if let Ok(port_str) = std::env::var("rain_GATEWAY_PORT").or_else(|_| std::env::var("PORT"))
-        {
-            if let Ok(port) = port_str.parse::<u16>() {
-                self.gateway.port = port;
-            }
-        }
-
-        // Gateway host: rain_GATEWAY_HOST or HOST
-        if let Ok(host) = std::env::var("rain_GATEWAY_HOST").or_else(|_| std::env::var("HOST")) {
-            if !host.is_empty() {
-                self.gateway.host = host;
-            }
-        }
-
-        // Allow public bind: rain_ALLOW_PUBLIC_BIND
-        if let Ok(val) = std::env::var("rain_ALLOW_PUBLIC_BIND") {
-            self.gateway.allow_public_bind = val == "1" || val.eq_ignore_ascii_case("true");
-        }
-
-        // Temperature: rain_TEMPERATURE
-        if let Ok(temp_str) = std::env::var("rain_TEMPERATURE") {
-            match temp_str.parse::<f64>() {
-                Ok(temp) if TEMPERATURE_RANGE.contains(&temp) => {
-                    self.default_temperature = temp;
-                }
-                Ok(temp) => {
-                    tracing::warn!(
-                        "Ignoring rain_TEMPERATURE={temp}: \
-                         value out of range (expected {}..={})",
-                        TEMPERATURE_RANGE.start(),
-                        TEMPERATURE_RANGE.end()
-                    );
-                }
-                Err(_) => {
-                    tracing::warn!("Ignoring rain_TEMPERATURE={temp_str:?}: not a valid number");
-                }
-            }
-        }
-
-        // Reasoning override: rain_REASONING_ENABLED or REASONING_ENABLED
-        if let Ok(flag) =
-            std::env::var("rain_REASONING_ENABLED").or_else(|_| std::env::var("REASONING_ENABLED"))
-        {
-            let normalized = flag.trim().to_ascii_lowercase();
-            match normalized.as_str() {
-                "1" | "true" | "yes" | "on" => self.runtime.reasoning_enabled = Some(true),
-                "0" | "false" | "no" | "off" => self.runtime.reasoning_enabled = Some(false),
-                _ => {}
-            }
-        }
-
-        if let Ok(raw) = std::env::var("rain_REASONING_EFFORT")
-            .or_else(|_| std::env::var("REASONING_EFFORT"))
-            .or_else(|_| std::env::var("rain_CODEX_REASONING_EFFORT"))
-        {
-            match normalize_reasoning_effort(&raw) {
-                Ok(effort) => self.runtime.reasoning_effort = Some(effort),
-                Err(message) => tracing::warn!("Ignoring reasoning effort env override: {message}"),
-            }
-        }
-
-        // Web search enabled: rain_WEB_SEARCH_ENABLED or WEB_SEARCH_ENABLED
-        if let Ok(enabled) = std::env::var("rain_WEB_SEARCH_ENABLED")
-            .or_else(|_| std::env::var("WEB_SEARCH_ENABLED"))
-        {
-            self.web_search.enabled = enabled == "1" || enabled.eq_ignore_ascii_case("true");
-        }
-
-        // Web search provider: rain_WEB_SEARCH_PROVIDER or WEB_SEARCH_PROVIDER
-        if let Ok(provider) = std::env::var("rain_WEB_SEARCH_PROVIDER")
-            .or_else(|_| std::env::var("WEB_SEARCH_PROVIDER"))
-        {
-            let provider = provider.trim();
-            if !provider.is_empty() {
-                self.web_search.provider = provider.to_string();
-            }
-        }
-
-        // Brave API key: rain_BRAVE_API_KEY or BRAVE_API_KEY
-        if let Ok(api_key) =
-            std::env::var("rain_BRAVE_API_KEY").or_else(|_| std::env::var("BRAVE_API_KEY"))
-        {
-            let api_key = api_key.trim();
-            if !api_key.is_empty() {
-                self.web_search.brave_api_key = Some(api_key.to_string());
-            }
-        }
-
-        // Web search max results: rain_WEB_SEARCH_MAX_RESULTS or WEB_SEARCH_MAX_RESULTS
-        if let Ok(max_results) = std::env::var("rain_WEB_SEARCH_MAX_RESULTS")
-            .or_else(|_| std::env::var("WEB_SEARCH_MAX_RESULTS"))
-        {
-            if let Ok(max_results) = max_results.parse::<usize>() {
-                if (1..=10).contains(&max_results) {
-                    self.web_search.max_results = max_results;
-                }
-            }
-        }
-
-        // Web search timeout: rain_WEB_SEARCH_TIMEOUT_SECS or WEB_SEARCH_TIMEOUT_SECS
-        if let Ok(timeout_secs) = std::env::var("rain_WEB_SEARCH_TIMEOUT_SECS")
-            .or_else(|_| std::env::var("WEB_SEARCH_TIMEOUT_SECS"))
-        {
-            if let Ok(timeout_secs) = timeout_secs.parse::<u64>() {
-                if timeout_secs > 0 {
-                    self.web_search.timeout_secs = timeout_secs;
-                }
-            }
-        }
-
-        // Storage provider key (optional backend override): rain_STORAGE_PROVIDER
-        if let Ok(provider) = std::env::var("rain_STORAGE_PROVIDER") {
-            let provider = provider.trim();
-            if !provider.is_empty() {
-                self.storage.provider.config.provider = provider.to_string();
-            }
-        }
-
-        // Storage connection URL (for remote backends): rain_STORAGE_DB_URL
-        if let Ok(db_url) = std::env::var("rain_STORAGE_DB_URL") {
-            let db_url = db_url.trim();
-            if !db_url.is_empty() {
-                self.storage.provider.config.db_url = Some(db_url.to_string());
-            }
-        }
-
-        // Storage connect timeout: rain_STORAGE_CONNECT_TIMEOUT_SECS
-        if let Ok(timeout_secs) = std::env::var("rain_STORAGE_CONNECT_TIMEOUT_SECS") {
-            if let Ok(timeout_secs) = timeout_secs.parse::<u64>() {
-                if timeout_secs > 0 {
-                    self.storage.provider.config.connect_timeout_secs = Some(timeout_secs);
-                }
-            }
-        }
-        // Proxy enabled flag: rain_PROXY_ENABLED
-        let explicit_proxy_enabled = std::env::var("rain_PROXY_ENABLED")
-            .ok()
-            .as_deref()
-            .and_then(parse_proxy_enabled);
-        if let Some(enabled) = explicit_proxy_enabled {
-            self.proxy.enabled = enabled;
-        }
-
-        // Proxy URLs: rain_* wins, then generic *PROXY vars.
-        let mut proxy_url_overridden = false;
-        if let Ok(proxy_url) =
-            std::env::var("rain_HTTP_PROXY").or_else(|_| std::env::var("HTTP_PROXY"))
-        {
-            self.proxy.http_proxy = normalize_proxy_url_option(Some(&proxy_url));
-            proxy_url_overridden = true;
-        }
-        if let Ok(proxy_url) =
-            std::env::var("rain_HTTPS_PROXY").or_else(|_| std::env::var("HTTPS_PROXY"))
-        {
-            self.proxy.https_proxy = normalize_proxy_url_option(Some(&proxy_url));
-            proxy_url_overridden = true;
-        }
-        if let Ok(proxy_url) =
-            std::env::var("rain_ALL_PROXY").or_else(|_| std::env::var("ALL_PROXY"))
-        {
-            self.proxy.all_proxy = normalize_proxy_url_option(Some(&proxy_url));
-            proxy_url_overridden = true;
-        }
-        if let Ok(no_proxy) = std::env::var("rain_NO_PROXY").or_else(|_| std::env::var("NO_PROXY"))
-        {
-            self.proxy.no_proxy = normalize_no_proxy_list(vec![no_proxy]);
-        }
-
-        if explicit_proxy_enabled.is_none()
-            && proxy_url_overridden
-            && self.proxy.has_any_proxy_url()
-        {
-            self.proxy.enabled = true;
-        }
-
-        // Proxy scope and service selectors.
-        if let Ok(scope_raw) = std::env::var("rain_PROXY_SCOPE") {
-            if let Some(scope) = parse_proxy_scope(&scope_raw) {
-                self.proxy.scope = scope;
-            } else {
-                tracing::warn!(
-                    scope = %scope_raw,
-                    "Ignoring invalid rain_PROXY_SCOPE (valid: environment|R.A.I.N.|services)"
-                );
-            }
-        }
-
-        if let Ok(services_raw) = std::env::var("rain_PROXY_SERVICES") {
-            self.proxy.services = normalize_service_list(vec![services_raw]);
-        }
-
-        if let Err(error) = self.proxy.validate() {
-            tracing::warn!("Invalid proxy configuration ignored: {error}");
-            self.proxy.enabled = false;
-        }
-
-        if self.proxy.enabled && self.proxy.scope == ProxyScope::Environment {
-            self.proxy.apply_to_process_env();
-        }
-
-        set_runtime_proxy_config(self.proxy.clone());
-
-        if self.conversational_ai.enabled {
-            tracing::warn!(
-                "conversational_ai.enabled = true but conversational AI features are not yet \
-                 implemented; this section is reserved for future use and will be ignored"
-            );
-        }
-    }
-
-    async fn resolve_config_path_for_save(&self) -> Result<PathBuf> {
-        if self
-            .config_path
-            .parent()
-            .is_some_and(|parent| !parent.as_os_str().is_empty())
-        {
-            return Ok(self.config_path.clone());
-        }
-
-        let (default_rain_dir, default_workspace_dir) = default_config_and_workspace_dirs()?;
-        let (rain_dir, _workspace_dir, source) =
-            resolve_runtime_config_dirs(&default_rain_dir, &default_workspace_dir).await?;
-        let file_name = self
-            .config_path
-            .file_name()
-            .filter(|name| !name.is_empty())
-            .unwrap_or_else(|| std::ffi::OsStr::new("config.toml"));
-        let resolved = rain_dir.join(file_name);
-        tracing::warn!(
-            path = %self.config_path.display(),
-            resolved = %resolved.display(),
-            source = source.as_str(),
-            "Config path missing parent directory; resolving from runtime environment"
-        );
-        Ok(resolved)
-    }
-
-    pub async fn save(&self) -> Result<()> {
-        // Encrypt secrets before serialization
-        let mut config_to_save = self.clone();
-        let config_path = self.resolve_config_path_for_save().await?;
-        let rain_dir = config_path
-            .parent()
-            .context("Config path must have a parent directory")?;
-        let store = crate::security::SecretStore::new(rain_dir, self.secrets.encrypt);
-
-        encrypt_optional_secret(&store, &mut config_to_save.api_key, "config.api_key")?;
-        encrypt_optional_secret(
-            &store,
-            &mut config_to_save.composio.api_key,
-            "config.composio.api_key",
-        )?;
-        if let Some(ref mut pinggy) = config_to_save.tunnel.pinggy {
-            encrypt_optional_secret(&store, &mut pinggy.token, "config.tunnel.pinggy.token")?;
-        }
-        encrypt_optional_secret(
-            &store,
-            &mut config_to_save.microsoft365.client_secret,
-            "config.microsoft365.client_secret",
-        )?;
-
-        encrypt_optional_secret(
-            &store,
-            &mut config_to_save.browser.computer_use.api_key,
-            "config.browser.computer_use.api_key",
-        )?;
-
-        encrypt_optional_secret(
-            &store,
-            &mut config_to_save.web_search.brave_api_key,
-            "config.web_search.brave_api_key",
-        )?;
-
-        encrypt_optional_secret(
-            &store,
-            &mut config_to_save.storage.provider.config.db_url,
-            "config.storage.provider.config.db_url",
-        )?;
-
-        for agent in config_to_save.agents.values_mut() {
-            encrypt_optional_secret(&store, &mut agent.api_key, "config.agents.*.api_key")?;
-        }
-
-        // Encrypt TTS provider API keys
-        if let Some(ref mut openai) = config_to_save.tts.openai {
-            encrypt_optional_secret(&store, &mut openai.api_key, "config.tts.openai.api_key")?;
-        }
-        if let Some(ref mut elevenlabs) = config_to_save.tts.elevenlabs {
-            encrypt_optional_secret(
-                &store,
-                &mut elevenlabs.api_key,
-                "config.tts.elevenlabs.api_key",
-            )?;
-        }
-        if let Some(ref mut google) = config_to_save.tts.google {
-            encrypt_optional_secret(&store, &mut google.api_key, "config.tts.google.api_key")?;
-        }
-
-        // Encrypt nested STT provider API keys
-        encrypt_optional_secret(
-            &store,
-            &mut config_to_save.transcription.api_key,
-            "config.transcription.api_key",
-        )?;
-        if let Some(ref mut openai) = config_to_save.transcription.openai {
-            encrypt_optional_secret(
-                &store,
-                &mut openai.api_key,
-                "config.transcription.openai.api_key",
-            )?;
-        }
-        if let Some(ref mut deepgram) = config_to_save.transcription.deepgram {
-            encrypt_optional_secret(
-                &store,
-                &mut deepgram.api_key,
-                "config.transcription.deepgram.api_key",
-            )?;
-        }
-        if let Some(ref mut assemblyai) = config_to_save.transcription.assemblyai {
-            encrypt_optional_secret(
-                &store,
-                &mut assemblyai.api_key,
-                "config.transcription.assemblyai.api_key",
-            )?;
-        }
-        if let Some(ref mut google) = config_to_save.transcription.google {
-            encrypt_optional_secret(
-                &store,
-                &mut google.api_key,
-                "config.transcription.google.api_key",
-            )?;
-        }
-
-        #[cfg(feature = "channel-nostr")]
-        if let Some(ref mut ns) = config_to_save.channels_config.nostr {
-            encrypt_secret(
-                &store,
-                &mut ns.private_key,
-                "config.channels_config.nostr.private_key",
-            )?;
-        }
-        if let Some(ref mut fs) = config_to_save.channels_config.feishu {
-            encrypt_secret(
-                &store,
-                &mut fs.app_secret,
-                "config.channels_config.feishu.app_secret",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut fs.encrypt_key,
-                "config.channels_config.feishu.encrypt_key",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut fs.verification_token,
-                "config.channels_config.feishu.verification_token",
-            )?;
-        }
-
-        // Encrypt channel secrets
-        if let Some(ref mut tg) = config_to_save.channels_config.telegram {
-            encrypt_secret(
-                &store,
-                &mut tg.bot_token,
-                "config.channels_config.telegram.bot_token",
-            )?;
-        }
-        if let Some(ref mut dc) = config_to_save.channels_config.discord {
-            encrypt_secret(
-                &store,
-                &mut dc.bot_token,
-                "config.channels_config.discord.bot_token",
-            )?;
-        }
-        if let Some(ref mut sl) = config_to_save.channels_config.slack {
-            encrypt_secret(
-                &store,
-                &mut sl.bot_token,
-                "config.channels_config.slack.bot_token",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut sl.app_token,
-                "config.channels_config.slack.app_token",
-            )?;
-        }
-        if let Some(ref mut mm) = config_to_save.channels_config.mattermost {
-            encrypt_secret(
-                &store,
-                &mut mm.bot_token,
-                "config.channels_config.mattermost.bot_token",
-            )?;
-        }
-        if let Some(ref mut mx) = config_to_save.channels_config.matrix {
-            encrypt_secret(
-                &store,
-                &mut mx.access_token,
-                "config.channels_config.matrix.access_token",
-            )?;
-        }
-        if let Some(ref mut wa) = config_to_save.channels_config.whatsapp {
-            encrypt_optional_secret(
-                &store,
-                &mut wa.access_token,
-                "config.channels_config.whatsapp.access_token",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut wa.app_secret,
-                "config.channels_config.whatsapp.app_secret",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut wa.verify_token,
-                "config.channels_config.whatsapp.verify_token",
-            )?;
-        }
-        if let Some(ref mut lq) = config_to_save.channels_config.linq {
-            encrypt_secret(
-                &store,
-                &mut lq.api_token,
-                "config.channels_config.linq.api_token",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut lq.signing_secret,
-                "config.channels_config.linq.signing_secret",
-            )?;
-        }
-        if let Some(ref mut wt) = config_to_save.channels_config.wati {
-            encrypt_secret(
-                &store,
-                &mut wt.api_token,
-                "config.channels_config.wati.api_token",
-            )?;
-        }
-        if let Some(ref mut nc) = config_to_save.channels_config.nextcloud_talk {
-            encrypt_secret(
-                &store,
-                &mut nc.app_token,
-                "config.channels_config.nextcloud_talk.app_token",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut nc.webhook_secret,
-                "config.channels_config.nextcloud_talk.webhook_secret",
-            )?;
-        }
-        if let Some(ref mut em) = config_to_save.channels_config.email {
-            encrypt_secret(
-                &store,
-                &mut em.password,
-                "config.channels_config.email.password",
-            )?;
-        }
-        if let Some(ref mut irc) = config_to_save.channels_config.irc {
-            encrypt_optional_secret(
-                &store,
-                &mut irc.server_password,
-                "config.channels_config.irc.server_password",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut irc.nickserv_password,
-                "config.channels_config.irc.nickserv_password",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut irc.sasl_password,
-                "config.channels_config.irc.sasl_password",
-            )?;
-        }
-        if let Some(ref mut lk) = config_to_save.channels_config.lark {
-            encrypt_secret(
-                &store,
-                &mut lk.app_secret,
-                "config.channels_config.lark.app_secret",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut lk.encrypt_key,
-                "config.channels_config.lark.encrypt_key",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut lk.verification_token,
-                "config.channels_config.lark.verification_token",
-            )?;
-        }
-        if let Some(ref mut fs) = config_to_save.channels_config.feishu {
-            encrypt_secret(
-                &store,
-                &mut fs.app_secret,
-                "config.channels_config.feishu.app_secret",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut fs.encrypt_key,
-                "config.channels_config.feishu.encrypt_key",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut fs.verification_token,
-                "config.channels_config.feishu.verification_token",
-            )?;
-        }
-        if let Some(ref mut dt) = config_to_save.channels_config.dingtalk {
-            encrypt_secret(
-                &store,
-                &mut dt.client_secret,
-                "config.channels_config.dingtalk.client_secret",
-            )?;
-        }
-        if let Some(ref mut wc) = config_to_save.channels_config.wecom {
-            encrypt_secret(
-                &store,
-                &mut wc.webhook_key,
-                "config.channels_config.wecom.webhook_key",
-            )?;
-        }
-        if let Some(ref mut qq) = config_to_save.channels_config.qq {
-            encrypt_secret(
-                &store,
-                &mut qq.app_secret,
-                "config.channels_config.qq.app_secret",
-            )?;
-        }
-        if let Some(ref mut wh) = config_to_save.channels_config.webhook {
-            encrypt_optional_secret(
-                &store,
-                &mut wh.secret,
-                "config.channels_config.webhook.secret",
-            )?;
-        }
-        if let Some(ref mut ct) = config_to_save.channels_config.clawdtalk {
-            encrypt_secret(
-                &store,
-                &mut ct.api_key,
-                "config.channels_config.clawdtalk.api_key",
-            )?;
-            encrypt_optional_secret(
-                &store,
-                &mut ct.webhook_secret,
-                "config.channels_config.clawdtalk.webhook_secret",
-            )?;
-        }
-
-        // Encrypt gateway paired tokens
-        for token in &mut config_to_save.gateway.paired_tokens {
-            encrypt_secret(&store, token, "config.gateway.paired_tokens[]")?;
-        }
-
-        // Encrypt Nevis IAM secret
-        encrypt_optional_secret(
-            &store,
-            &mut config_to_save.security.nevis.client_secret,
-            "config.security.nevis.client_secret",
-        )?;
-
-        // Notion API key (top-level, not in ChannelsConfig)
-        if !config_to_save.notion.api_key.is_empty() {
-            encrypt_secret(
-                &store,
-                &mut config_to_save.notion.api_key,
-                "config.notion.api_key",
-            )?;
-        }
-
-        // Jira API token
-        if !config_to_save.jira.api_token.is_empty() {
-            encrypt_secret(
-                &store,
-                &mut config_to_save.jira.api_token,
-                "config.jira.api_token",
-            )?;
-        }
-
-        let toml_str =
-            toml::to_string_pretty(&config_to_save).context("Failed to serialize config")?;
-
-        let parent_dir = config_path
-            .parent()
-            .context("Config path must have a parent directory")?;
-
-        fs::create_dir_all(parent_dir).await.with_context(|| {
-            format!(
-                "Failed to create config directory: {}",
-                parent_dir.display()
-            )
-        })?;
-
-        let file_name = config_path
-            .file_name()
-            .and_then(|v| v.to_str())
-            .unwrap_or("config.toml");
-        let temp_path = parent_dir.join(format!(".{file_name}.tmp-{}", uuid::Uuid::new_v4()));
-        let backup_path = parent_dir.join(format!("{file_name}.bak"));
-
-        let mut temp_file = OpenOptions::new()
-            .create_new(true)
-            .write(true)
-            .open(&temp_path)
-            .await
-            .with_context(|| {
-                format!(
-                    "Failed to create temporary config file: {}",
-                    temp_path.display()
-                )
-            })?;
-        temp_file
-            .write_all(toml_str.as_bytes())
-            .await
-            .context("Failed to write temporary config contents")?;
-        temp_file
-            .sync_all()
-            .await
-            .context("Failed to fsync temporary config file")?;
-        drop(temp_file);
-
-        let had_existing_config = config_path.exists();
-        if had_existing_config {
-            fs::copy(&config_path, &backup_path)
-                .await
-                .with_context(|| {
-                    format!(
-                        "Failed to create config backup before atomic replace: {}",
-                        backup_path.display()
-                    )
-                })?;
-        }
-
-        if let Err(e) = fs::rename(&temp_path, &config_path).await {
-            let _ = fs::remove_file(&temp_path).await;
-            if had_existing_config && backup_path.exists() {
-                fs::copy(&backup_path, &config_path)
-                    .await
-                    .context("Failed to restore config backup")?;
-            }
-            anyhow::bail!("Failed to atomically replace config file: {e}");
-        }
-
-        #[cfg(unix)]
-        {
-            use std::{fs::Permissions, os::unix::fs::PermissionsExt};
-            if let Err(err) = fs::set_permissions(&config_path, Permissions::from_mode(0o600)).await
-            {
-                tracing::warn!(
-                    "Failed to harden config permissions to 0600 at {}: {}",
-                    config_path.display(),
-                    err
-                );
-            }
-        }
-
-        sync_directory(parent_dir).await?;
-
-        if had_existing_config {
-            let _ = fs::remove_file(&backup_path).await;
-        }
-
-        Ok(())
-    }
 }
 
 #[allow(clippy::unused_async)] // async needed on unix for tokio File I/O; no-op on other platforms
